When the user creates a new item to the list, it will now create a listener for the checkbox to be checked. When the checkbox is checkbox is clicked, the text next to it associated with the same todo is struck through (line through it) and when it is unchecked the text returns to normal. This now works when you add a new item to the list
KNOWN ISSUE: When you click a checkbox and cause a toDoItem to be completed (setting the text to be struckthrough), when you add a new item to the list, it will also be struck through for some reason. If you click another check box on a completed todo, any new todoItems will appear normally for some reason.